Refactor ConstraintComponents to Constraint and rename physics comps
