+++ cocos/physics/framework/components/colliders/capsule-collider.ts
@@ -11,7 +11,7 @@ import {
     tooltip,
     type,
 } from '../../../../core/data/class-decorator';
-import { Collider } from './collider-component';
+import { Collider } from './collider';
 import { ICapsuleShape } from '../../../spec/i-physics-shape';
 import { EDITOR, TEST } from 'internal:constants';
 import { EAxisDirection, EColliderType } from '../../physics-enum';
