+++ cocos/physics/framework/index.ts
@@ -5,23 +5,23 @@
 import { PhysicsSystem } from './physics-system';
 import { PhysicMaterial } from './assets/physic-material';
 import { PhysicsRayResult } from './physics-ray-result';
-import { BoxCollider } from './components/colliders/box-collider-component';
-import { Collider } from './components/colliders/collider-component';
-import { SphereCollider } from './components/colliders/sphere-collider-component';
-import { CapsuleCollider } from './components/colliders/capsule-collider-component';
-import { CylinderCollider } from './components/colliders/cylinder-collider-component';
-import { ConeCollider } from './components/colliders/cone-collider-component';
-import { MeshCollider } from './components/colliders/mesh-collider-component';
-import { RigidBody } from './components/rigid-body-component';
+import { BoxCollider } from './components/colliders/box-collider';
+import { Collider } from './components/colliders/collider';
+import { SphereCollider } from './components/colliders/sphere-collider';
+import { CapsuleCollider } from './components/colliders/capsule-collider';
+import { CylinderCollider } from './components/colliders/cylinder-collider';
+import { ConeCollider } from './components/colliders/cone-collider';
+import { MeshCollider } from './components/colliders/mesh-collider';
+import { RigidBody } from './components/rigid-body';
 import { ConstantForce } from './components/constant-force';
-import { TerrainCollider } from './components/colliders/terrain-collider-component';
-import { SimplexCollider } from './components/colliders/simplex-collider-component';
-import { PlaneCollider } from './components/colliders/plane-collider-component';
+import { TerrainCollider } from './components/colliders/terrain-collider';
+import { SimplexCollider } from './components/colliders/simplex-collider';
+import { PlaneCollider } from './components/colliders/plane-collider';
 
 // constraints
-import { Constraint } from './components/constraints/constraint-component';
-import { HingeConstraint } from './components/constraints/hinge-constraint-component';
-import { PointToPointConstraint } from './components/constraints/point-to-point-constraint-component';
+import { Constraint } from './components/constraints/constraint';
+import { HingeConstraint } from './components/constraints/hinge-constraint';
+import { PointToPointConstraint } from './components/constraints/point-to-point-constraint';
 
 export {
     PhysicsSystem,
